+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Add new method :meth:`pycov.CovFrame.mask_bed`.
 * Add new method :meth:`pycov.CovFrame.chr_prefix`.
 * Add new property ``contigs`` to :class:`pybed.BedFrame`.
+* Add new method :meth:`pybed.BedFrame.chr_prefix`.
 
 0.22.0 (2021-09-04)
 -------------------